+++ src/GSave.cxx
@@ -25,6 +25,7 @@
 //    http://www.fltk.org/str.php
 //
 
+#include <cstdint>
 #include <config.h>
 #include <fltk/x.h>
 namespace fltk {class Image;}
@@ -79,11 +80,11 @@ GSave::GSave() {
 # error
 #endif
   data[2] = fl_current_Image;
-  data[3] = (void*)((fl_clip_w<<16)+(fl_clip_h&0xffff));
+  data[3] = (void*)(uintptr_t)((fl_clip_w<<16)+(fl_clip_h&0xffff));
 }
 
 GSave::~GSave() {
-  unsigned v = (unsigned) (unsigned long)data[3];
+  unsigned v = (unsigned) (uintptr_t)data[3];
   fl_clip_w = v >> 16;
   fl_clip_h = v & 0xffff;
 #if USE_X11
